page_objects_model_selenium/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
asgiref==3.8.1
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.2.0
cssutils==2.11.1
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
dotenv==0.9.9
exceptiongroup==1.2.2
execnet==2.1.1
Faker==20.1.0
Flask==3.0.2
Flask-Cors==4.0.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.2
lxml==5.3.1
MarkupSafe==3.0.2
more-itertools==10.6.0
outcome==1.3.0.post0
packaging==24.2
playwright==1.50.0
pluggy==1.5.0
premailer==3.10.0
pyee==12.1.1
PySocks==1.7.1
pytest==7.4.3
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-playwright==0.4.3
pytest-rerunfailures==12.0
pytest-selenium==4.1.0
pytest-timeout==2.2.0
pytest-variables==3.0.0
pytest-xdist==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2025.1
requests==2.31.0
selenium==4.15.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tenacity==9.0.0
text-unidecode==1.3
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yagmail==0.15.293
